Filesystem: Moved toolchain-specific types into retarget.h
